requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
blessings==1.7
bpython==0.19
certifi==2020.6.20
chardet==3.0.4
curtsies==0.3.1
distlib==0.3.0
Django==2.2.9
django-bootstrap4==2.1.1
django-debug-toolbar==2.2
filelock==3.0.12
greenlet==0.4.16
idna==2.9
more-itertools==8.3.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.8.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.25
wcwidth==0.2.4
django-bootstrap-pagination==1.7.1
